+++ src/main/java/com/petboard/PetBoardApplication.java
@@ -1,0 +1,13 @@
+package com.petboard;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PetBoardApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PetBoardApplication.class, args);
+    }
+
+}
