+++ src/main/java/com/petboard/domain/user/dto/UserLoginRequestDto.java
@@ -1,0 +1,9 @@
+package com.petboard.domain.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UserLoginRequestDto {
+    private String email;
+    private String password;
+}
