+++ src/main/java/com/petboard/domain/user/dto/UserLoginResponseDto.java
@@ -1,0 +1,10 @@
+package com.petboard.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserLoginResponseDto {
+    private String token;
+}
