+++ src/main/java/com/petboard/domain/user/dto/UserResponseDto.java
@@ -1,0 +1,17 @@
+package com.petboard.domain.user.dto;
+
+import com.petboard.domain.user.entity.User;
+import lombok.Getter;
+
+@Getter
+public class UserResponseDto {
+    private final Long userId;
+    private final String email;
+    private final String nickname;
+
+    public UserResponseDto(User user) {
+        this.userId = user.getUserId();
+        this.email = user.getEmail();
+        this.nickname = user.getNickname();
+    }
+}
