+++ src/main/java/com/petboard/domain/user/dto/UserSignUpResponseDto.java
@@ -1,0 +1,13 @@
+package com.petboard.domain.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserSignUpResponseDto {
+
+    private Long userId;
+    private String email;
+    private String nickname;
+}
