+++ src/main/java/com/petboard/domain/user/entity/Grade.java
@@ -1,0 +1,19 @@
+package com.petboard.domain.user.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+public class Grade {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Byte gradeNo;
+
+    @Column(nullable = false, length = 10)
+    private String gradeName;
+}
