--- src/main/java/me/jungtaemin/petboard/PetBoardApplication.java
@@ -1,13 +1,0 @@
-package me.jungtaemin.petboard;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PetBoardApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PetBoardApplication.class, args);
-    }
-
-}
