--- src/main/java/me/jungtaemin/petboard/domain/user/dto/UserLoginRequestDto.java
@@ -1,9 +1,0 @@
-package me.jungtaemin.petboard.domain.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserLoginRequestDto {
-    private String email;
-    private String password;
-}
