--- src/main/java/me/jungtaemin/petboard/domain/user/dto/UserLoginResponseDto.java
@@ -1,10 +1,0 @@
-package me.jungtaemin.petboard.domain.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserLoginResponseDto {
-    private String token;
-}
