--- src/main/java/me/jungtaemin/petboard/domain/user/dto/UserSignUpRequestDto.java
@@ -1,15 +1,0 @@
-package me.jungtaemin.petboard.domain.user.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class UserSignUpRequestDto {
-
-    private String email;
-    private String password;
-    private String nickname;
-
-    private Long gradeId;
-}
