--- src/main/java/me/jungtaemin/petboard/domain/user/dto/UserSignUpResponseDto.java
@@ -1,13 +1,0 @@
-package me.jungtaemin.petboard.domain.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserSignUpResponseDto {
-
-    private Long userId;
-    private String email;
-    private String nickname;
-}
