--- src/main/java/me/jungtaemin/petboard/domain/user/entity/Grade.java
@@ -1,19 +1,0 @@
-package me.jungtaemin.petboard.domain.user.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class Grade {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Byte gradeNo;
-
-    @Column(nullable = false, length = 10)
-    private String gradeName;
-}
