--- src/main/java/me/jungtaemin/petboard/domain/user/entity/User.java
@@ -1,34 +1,0 @@
-package me.jungtaemin.petboard.domain.user.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "grade_id", nullable = false)
-    private Grade grade;
-
-    @Column(nullable = false, length = 100, unique = true)
-    private String email;
-
-    @Column(nullable = false, length = 255)
-    private String password;
-
-    @Column(nullable = false, length = 50)
-    private String nickname;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-}
