--- src/main/java/me/jungtaemin/petboard/domain/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package me.jungtaemin.petboard.domain.user.repository;
-
-import me.jungtaemin.petboard.domain.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
-    boolean existsByNickname(String nickname);
-    Optional<User> findByEmail(String email);
-}
