--- src/main/java/me/jungtaemin/petboard/global/config/SwaggerConfig.java
@@ -1,27 +1,0 @@
-package me.jungtaemin.petboard.global.config;
-
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@SecurityScheme(
-        name = "JWT",
-        type = SecuritySchemeType.HTTP,
-        scheme = "bearer",
-        bearerFormat = "JWT"
-)
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("PetBoard API")
-                        .version("1.0")
-                        .description("반려동물 커뮤니티 서비스 API 명세서"));
-    }
-}
