--- src/main/java/me/jungtaemin/petboard/global/response/ApiResponse.java
@@ -1,31 +1,0 @@
-package me.jungtaemin.petboard.global.response;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ApiResponse<T> {
-
-    private T data;
-    private String message;
-    private int statusCode;
-
-    private ApiResponse(T data, String message, int statusCode) {
-        this.data = data;
-        this.message = message;
-        this.statusCode = statusCode;
-    }
-
-    public static <T> ApiResponse<T> success(T data) {
-        return new ApiResponse<>(data, "요청 성공", 200);
-    }
-
-    public static <T> ApiResponse<T> success(T data, String message) {
-        return new ApiResponse<>(data, message, 200);
-    }
-
-    public static <T> ApiResponse<T> error(String message, int statusCode) {
-        return new ApiResponse<>(null, message, statusCode);
-    }
-}
