+++ src/test/java/com/jungtaemin/petboard/PetBoardApplicationTests.java
@@ -1,0 +1,13 @@
+package com.jungtaemin.petboard;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PetBoardApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
