--- src/test/java/me/jungtaemin/petboard/PetBoardApplicationTests.java
@@ -1,13 +1,0 @@
-package me.jungtaemin.petboard;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PetBoardApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
